backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-environ==0.11.2
django-redis==5.4.0
django-celery-beat==2.6.0
celery==5.3.6
redis==5.0.3
psycopg2-binary==2.9.9
pandas==2.2.2
pyarrow==16.1.0
python-json-logger==2.0.7
django-cors-headers==4.3.1
gunicorn==22.0.0
pytest==8.2.2
pytest-django==4.8.0
pytest-mock==3.14.0
factory-boy==3.3.0
ruff==0.11.0